[feat] Export text with styles (#3)
* [feat] Fix export when it has style with not styled text

* [fix] Fix error of removing space after formatting

* [feat] Add tests to formatting at the end of sentence

* [feat] Fix italic and underline export

This commit fixes the same problem as the previous one. The error was making a
text with style to be placed after a text without style

* [feat] Fix error of text content without <Text>

* [fix] Fix error when underline texts wrap not formatting

This commit fixes an error that happens when we have a text that in both
sides have underline applied and in the middle has no text.
E.g. _aaa_ bbbb _ccc_

* [feat] Allow export '*_' to FDX

With this commit when the text has '*_' chars escaped we export them to FDX
without translating them as part of a style delimiter

* [test] DRY tests

* [test] Improve readbility of tests

* [test] Add test to scenario that does not work

* [refactor] Address code review suggestions