[fix] Don't create dummy headings on top of FDX script (#4)
When we make sure all text content is wrapped on a `<Text>`, we cannot put script elements (`<Paragraph Type=...>`) inside a `<Text>` on the `<Content>`, otherwise Final Draft understands that structure as multiple empty headings on top of the script. #buggyFD

So instead of an FDX file like this:
```
<FinalDraft DocumentType="Script" Template="No" Version="1">
  <Content>
    <Text>
      <Paragraph Type="Action"><Text>...</Text></Paragraph>
    </Text>
  </Content>
</FinalDraft>
```
, we need to have a file like this:
```
<FinalDraft DocumentType="Script" Template="No" Version="1">
  <Content>
    <Paragraph Type="Action"><Text>...</Text></Paragraph>
  </Content>
</FinalDraft>
```

Fix [this bug](https://trello.com/c/re4ST6Ld/1737).